Patch 1 applied from luiscubal
git-svn-id: http://flashdevelop.googlecode.com/svn/trunk@163 2354be2f-a62e-0410-afb7-89360a396392

+++ FD3/External/Plugins/FdbPlugin/PluginMain.cs
@@ -348,7 +348,7 @@ namespace FdbPlugin
         private void fdbWrapper_PrintEvent(Object sender, PrintArg e)
         {
             String name = e.valname;
-            if (e.printtype == (int)PrintType.LOCAL)
+            if (e.printtype == PrintType.LOCAL)
             {
                 pluginUI.TreeControl.Invoke((MethodInvoker)delegate()
                 {
